Exam: Explain wrong answer and grade essay question

+++ src/main/java/com/cabybara/prolearningplatform/dto/response/exam/ExamAnswerResultDto.java
@@ -6,6 +6,6 @@ public record ExamAnswerResultDto(
         String essayAnswer,
         Boolean isCorrect,
         String expectedAnswer,
-        Integer earnedPoints,
+        Double earnedPoints,
         String feedback
 ) {}
