+++ src/main/java/com/cabybara/prolearningplatform/dto/response/exam/ExamAttemptDto.java
@@ -12,6 +12,6 @@ public record ExamAttemptDto(
         LocalDateTime startedAt,
         LocalDateTime deadlineAt,
         LocalDateTime submittedAt,
-        BigDecimal score,
+        Double score,
         Integer totalPoints
 ) {}
